+++ src/components/Datatable/Table/TableHeadCell/index.ts
@@ -1,0 +1,1 @@
+export { default as TableHeadCell } from './TableHeadCell';